We are now able to instantiate a CharacterCreature
Added a Main class to instantiate a character.
Refactored the SlotManager so it does not manage buffs anymore, that
functionality has been moved to BuffManager.
Removed the race interface and made CharacterCreature just take a
humanoid, all the parts of Race have been moved to Humanoid.
Added a source id to each RWOModifier so the source RWO can be
identified.
Added and changed other various parts just about everywhere.