Combat Simulator v0.1
Not sure if it will run when pulled, try configuring eclipse to add
miglayout15-swing.jar to the build path if you can't start the combat
simulator. Eclipse automatically added it to the project director when I
began using the mig layout from swing.

Combat Simulator v0.1 features
Can add any weapon, armor, shield to the battle field.
Can add any Character creature of any race, class, and level.

Still need to implement basically everything else about combat simulator
such as combat rounds, controlling and viewing RWO's on the field,
attacks, and other actions.

Simple to implement is adding a piece of equipment to a creature as its
in the code ready to be used from the menu. I just forgot to add it in.

Also added a few enums that list the different races and classes so I
could build a few factories and make it easy to create new objects.

Hurray we finally have something that runs and looks like a program.