Lots of changes in relation to the CombatSimulator
The biggest change is the addition of the status menu which can be used
to view the contents of cells.

Also added a bunch of other changes here and there as needed.
Not all but a few are:
New EnumUtil class which contains a method that will alphabetically sort
enums.
Moved Cell and Coordinate to the pathfinder.world package.
Fixed an error in Inventory so it would return the list of inventory
instead of an empty list.
Changed a few methods to return Collections instead of Lists to make
them more general, I don't know if that's better though.
Added a public String getProperties() method to RealWorldObjects meant
to be overriden by all its subclasses to their properties as a list of
strings, useful for the status menu.
Made the New Simulation button work on the CombatSimulator, though I
have a feeling it could definitely be optimized.